sketchcode/requirements.txt
gunicorn>=19.7.1
six==1.11.0

Flask==0.11.1
Flask-Cors==1.8.0
Pillow==2.6.1
tinys3==0.1.11
h5py==2.7.1
keras==2.1.3
numpy>=1.11
scipy>=0.18
scikit-learn
ipython<6.0
tensorflow>=1.0
